+++ package.json
@@ -18,6 +18,7 @@
     "axios": "^1.9.0",
     "cpx": "^1.5.0",
     "express": "^5.1.0",
+    "express-file-routing": "^3.1.0",
     "jsonwebtoken": "^9.0.2",
     "nostr-tools": "^2.12.0",
     "rimraf": "^6.0.1",
