+++ src/app.ts
@@ -2,27 +2,23 @@ import express, { Application } from "express";
 import { logger } from "./lib/middlewares/logger.middleware.js";
 import { errorHandler } from "./lib/middlewares/error-handler.middleware.js";
 
-import healthRouter from "./routes/health.routes.js";
-import authRouter from "./routes/auth.routes.js";
-import usersRouter from "./routes/user.routes.js";
-import eventsRouter from "./routes/events.routes.js";
-import checkinRouter from "./routes/checkin.routes.js";
+import { router } from "express-file-routing";
+import path, { dirname } from "path";
+import { fileURLToPath } from "url";
 
-const createApp = (): Application => {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const createApp = async (): Promise<Application> => {
   const app: Application = express();
 
   // Middleware
   app.use(express.json());
   app.use(logger);
 
-  // Routes
-  app.use("/health", healthRouter);
-  app.use("/auth", authRouter);
-  app.use("/users", usersRouter);
-  app.use("/events", eventsRouter);
-  app.use("/checkin", checkinRouter);
+  app.use("/", await router({ directory: path.join(__dirname, "routes") }));
 
-  // 404 handler for undefined routes
+  // // 404 handler for undefined routes
   app.use((req, res) => {
     res.status(404).json({
       status: "error",
@@ -31,7 +27,7 @@ const createApp = (): Application => {
     });
   });
 
-  // Error handling middleware (should be last)
+  // // Error handling middleware (should be last)
   app.use(errorHandler);
 
   return app;
