+++ src/index.ts
@@ -3,12 +3,11 @@ import { config } from "./lib/env.js";
 
 const startServer = async (): Promise<void> => {
   try {
-    const app = createApp();
+    const app = await createApp();
     const PORT = config.port;
 
     app.listen(PORT, () => {
       console.log(`🚀 Server started successfully in ${config.nodeEnv} mode`);
-      console.log(`🔗 Health endpoint: http://localhost:${PORT}/health`);
       console.log(`⏱️ Server time: ${new Date().toISOString()}`);
     });
   } catch (error) {
