--- src/routes/tickets.routes.ts
@@ -1,53 +1,0 @@
-import express, { Router } from "express";
-import {
-  createTicket,
-  deleteTicket,
-  getAdminTickets,
-  getPublicTickets,
-  updateTicket,
-} from "../controllers/ticket.controller.js";
-import { requireAuth } from "../lib/middlewares/require-auth.middleware.js";
-import { requireEventRole } from "../lib/middlewares/required-event-role.middleware.js";
-import { validate } from "../lib/middlewares/validate.middleware.js";
-import {
-  CreateTicketSchema,
-  UpdateTicketSchema,
-} from "../lib/validators/ticket.schema.js";
-
-const router: Router = express.Router();
-
-// Public listing (no auth)
-router.get("/", getPublicTickets);
-
-// Admin listing (auth)
-router.get(
-  "/admin",
-  requireAuth,
-  requireEventRole(["OWNER", "MODERATOR"]),
-  getAdminTickets,
-);
-
-router.post(
-  "/",
-  requireAuth,
-  requireEventRole(["OWNER", "MODERATOR"]),
-  validate(CreateTicketSchema),
-  createTicket,
-);
-
-router.patch(
-  "/:ticketId",
-  requireAuth,
-  requireEventRole(["OWNER", "MODERATOR"]),
-  validate(UpdateTicketSchema),
-  updateTicket,
-);
-
-router.delete(
-  "/:ticketId",
-  requireAuth,
-  requireEventRole(["OWNER", "MODERATOR"]),
-  deleteTicket,
-);
-
-export default router;
